requirements.txt
Flask
Pillow
pillow-heif
Wand
gunicorn